+++ docs/changes.rst
@@ -7,6 +7,7 @@ Change Log
   * created :func:`pysftp.walktree` for walking local directories
   * added param recurse to :meth:`.pysftp.Connection.walktree` to allow it to do another trick
   * created :meth:`.put_d` to put the contents of a local directory to a remote one
+  * created a context manager chdir method, :meth:`.cd`
 
 * 0.2.7 (current, released 2014-05-24)